Completed adding and reming players to dom + local storage through form.

+++ src/components/PlayerList.js
@@ -13,9 +13,15 @@ class PlayerList {
     Storage.savePlayer(newPlayer);
     this._displayPlayerList();
   }
+
   removePlayerFromList(id) {
-    Storage.removePlayer(id);
-    this._displayPlayerList();
+    const index = this.list.findIndex((player) => player.id === id);
+    if (index !== -1) {
+      this.list.splice(index, 1);
+      Storage.removePlayer(id);
+      this._displayPlayerList();
+      console.log('Player removed');
+    }
   }
 
   _displayPlayerList() {
@@ -25,18 +31,20 @@ class PlayerList {
       const PlayerEl = document.createElement('div');
       PlayerEl.setAttribute('data-id', `${player.id}`);
       PlayerEl.classList.add(
+        'player',
         'border',
         'flex',
         'justify-around',
         'items-center',
-        'bg-white'
+        'bg-white',
+        'p-1'
       );
-      PlayerEl.innerHTML = `${player.name} <i class="fa-solid fa-user-xmark  text-red-600"></i>`;
+      PlayerEl.innerHTML = `<span>${player.name}</span> <button class="delete-button px-3 py-1"><i class="fa-solid fa-xmark fa-lg text-red-600"></i></button>`;
       PlayerListEl.appendChild(PlayerEl);
     });
+    console.log('Players loaded...');
   }
 
-  //   @todo removePlayerFromList(id)
   //   @todo shuffleListAll()
   //   @todo shuffleListBySex()
 }
